# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CookieLoginRequest": {
        "properties": {
          "cookies": {
            "description": "Cookie字符串，格式: 'INCO=xxx; SESSION=yyy'",
            "example": "INCO=abc123; SESSION=xyz789",
            "title": "Cookies",
            "type": "string"
          }
        },
        "required": [
          "cookies"
        ],
        "title": "CookieLoginRequest",
        "type": "object"
      },
      "CookieLoginResponse": {
        "example": {
          "status": "success",
          "student_id": "41234567",
          "student_name": "张三"
        },
        "properties": {
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "会话ID",
            "title": "Session Id"
          },
          "status": {
            "description": "登录状态",
            "title": "Status",
            "type": "string"
          },
          "student_id": {
            "description": "学号",
            "title": "Student Id",
            "type": "string"
          },
          "student_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "学生姓名",
            "title": "Student Name"
          }
        },
        "required": [
          "status",
          "student_id"
        ],
        "title": "CookieLoginResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QRInitResponse": {
        "example": {
          "qr_image": "data:image/png;base64,iVBORw0KGgo...",
          "session_id": "abc123xyz"
        },
        "properties": {
          "qr_image": {
            "description": "Base64编码的二维码图片，格式为data:image/png;base64,...",
            "title": "Qr Image",
            "type": "string"
          },
          "session_id": {
            "description": "会话ID，用于后续轮询",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "qr_image"
        ],
        "title": "QRInitResponse",
        "type": "object"
      },
...
      "SimpleResponse": {
        "example": {
          "status": "ok"
        },
        "properties": {
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "更新后的会话ID",
            "title": "Session Id"
          },
          "status": {
            "description": "操作状态",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "SimpleResponse",
        "type": "object"
      },
      "SmsRequest": {
        "properties": {
          "phone": {
            "description": "手机号码",
            "example": "13800138000",
            "title": "Phone",
            "type": "string"
          }
        },
        "required": [
          "phone"
        ],
        "title": "SmsRequest",
        "type": "object"
      },
      "SmsVerifyRequest": {
        "properties": {
          "code": {
            "description": "短信验证码",
            "example": "123456",
            "title": "Code",
            "type": "string"
          },
          "phone": {
            "description": "手机号码",
            "example": "13800138000",
            "title": "Phone",
            "type": "string"
          }
        },
        "required": [
          "phone",
          "code"
        ],
        "title": "SmsVerifyRequest",
        "type": "object"
      },
      "StatusResponse": {
        "example": {
          "authenticated": true,
          "state": "active"
        },
        "properties": {
          "authenticated": {
            "description": "是否已认证",
            "title": "Authenticated",
            "type": "boolean"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "认证状态：init/qr_ready/active等",
            "title": "State"
          }
        },
        "required": [
          "authenticated"
        ],
        "title": "StatusResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "USTB Manager",
      "url": "https://github.com/yourusername/ustb-manager"
    },
    "description": "\n## USTB Manager - 北科大教务管理系统API\n\n这是一个用于访问北京科技大学教务系统的API服务，提供以下功能：\n\n### 🔐 认证模块 (Authentication)\n- **二维码登录**: 使用微信扫码快速登录\n- **短信验证码登录**: 通过手机号接收验证码登录\n- **Cookie登录**: 使用浏览器Cookie直接登录（高级用户）\n- **会话管理**: 自动管理用户会话，支持后端重启后恢复\n\n### 📊 成绩查询模块 (Grades)\n- **成绩列表**: 支持分页、筛选的成绩查询\n- **GPA计算**: 自动计算总GPA和加权GPA\n- **学生信息**: 获取学生基本信息和完整档案\n- **培养方案**: 查看专业培养方案和课程要求\n- **学业进度**: 查询必修课完成情况和学分统计\n\n### 🔒 安全特性\n- HttpOnly Cookie保护\n- Session自动过期和清理\n- Cookie有效性验证\n- 支持CORS跨域访问\n\n### 📝 使用说明\n1. 首先调用认证接口（QR/SMS/Cookie）完成登录\n2. 登录成功后会自动设置session cookie\n3. 后续请求会自动携带cookie进行认证\n4. 可以调用 `/api/auth/status` 检查登录状态\n5. 使用 `/api/auth/logout` 退出登录\n\n### 🌐 环境信息\n- 开发环境: http://localhost:8000\n- API前缀: `/api`\n- 文档地址: `/docs` (Swagger UI) 或 `/redoc` (ReDoc)\n\n### 📞 技术支持\n如有问题，请查看各接口的详细文档说明。\n    ",
    "license": {
      "name": "MIT License"
    },
    "title": "USTB Manager API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/cookie/login": {
      "post": {
        "description": "## 业务说明\n使用从浏览器复制的USTB教务系统Cookie直接登录，适合高级用户快速登录。\n\n## 使用流程\n1. 用户在浏览器中登录USTB教务系统\n2. 从浏览器开发者工具复制Cookie字符串\n3. 粘贴到登录表单并提交\n4. 系统验证Cookie有效性\n5. 验证成功后创建session\n\n## Cookie格式\n```\nINCO=abc123; SESSION=xyz789\n```\n\n## 验证机制\n- 系统会使用提供的Cookie访问教务系统API\n- 尝试获取学生信息以验证Cookie有效性\n- 如果能成功获取学号和姓名，则认为Cookie有效\n\n## 注意事项\n- Cookie必须包含USTB教务系统的认证信息\n- Cookie过期会返回401错误\n- 验证成功后会保存Cookie到本地，支持后端重启后恢复\n- 会自动设置session cookie\n\n## 安全提示\n- Cookie包含敏感信息，请勿分享给他人\n- 建议定期更换Cookie",
        "operationId": "cookie_login_api_auth_cookie_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CookieLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CookieLoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "使用Cookie登录",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "description": "## 业务说明\n退出登录，清除session和cookie。\n\n## 使用流程\n1. 用户点击退出按钮\n2. 调用此接口\n3. 前端跳转到登录页\n\n## 注意事项\n- 会清除服务器端的session数据\n- 会删除客户端的cookie\n- 即使没有有效session也会返回成功",
        "operationId": "logout_api_auth_logout_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "退出登录",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/qr/complete": {
      "post": {
        "description": "## 业务说明\n完成二维码登录流程，刷新session并设置长期cookie。\n\n## 使用流程\n1. 在 `/qr/status` 收到success状态后调用此接口\n2. 服务器会轮换session ID（安全措施）\n3. 设置新的session cookie\n\n## 注意事项\n- 必须在认证成功后调用\n- 会自动更新cookie，前端无需处理\n- 调用后即可访问需要认证的API",
        "operationId": "qr_complete_api_auth_qr_complete_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "完成二维码登录",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/qr/init": {
      "post": {
        "description": "## 业务说明\n初始化微信扫码登录流程，生成登录二维码。\n\n## 使用流程\n1. 调用此接口获取二维码图片和session_id\n2. 前端展示二维码供用户扫描\n3. 使用返回的session_id调用 `/qr/status` 轮询登录状态\n4. 登录成功后调用 `/qr/complete` 完成认证\n\n## 注意事项\n- 二维码有效期约5分钟\n- 接口会自动设置session cookie (ustb_sid)\n- 二维码图片为Base64编码，可直接用于img标签的src属性\n- 如果已有有效的authenticated session，会复用该session",
        "operationId": "qr_init_api_auth_qr_init_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QRInitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "初始化二维码登录",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/qr/poll": {
      "get": {
        "description": "## 业务说明\n非SSE方式轮询二维码扫码状态，适用于不支持SSE的客户端（如微信小程序）。\n\n## 使用流程\n1. 调用 `/qr/init` 获取二维码后\n2. 每隔2秒调用此接口轮询状态\n3. 收到 success 后调用 `/qr/complete` 完成登录\n\n## 状态说明\n- `waiting`: 等待扫码\n- `scanned`: 已扫码，等待确认\n- `success`: 登录成功\n- `expired`: 二维码已过期",
        "operationId": "qr_poll_api_auth_qr_poll_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "轮询二维码登录状态（非SSE）",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/qr/status": {
      "get": {
        "description": "## 业务说明\n使用Server-Sent Events (SSE)实时推送二维码扫码状态。\n\n## 使用流程\n1. 前端使用EventSource连接此接口\n2. 服务器会持续推送状态更新\n3. 收到success状态后，调用 `/qr/complete` 完成登录\n\n## 状态说明\n- `waiting`: 等待扫码\n- `scanned`: 已扫码，等待确认\n- `success`: 登录成功\n- `expired`: 二维码已过期\n- `error`: 发生错误\n\n## 注意事项\n- 需要携带 ustb_sid cookie\n- 使用SSE协议，前端需使用EventSource API\n- 连接会在收到终态(success/expired/error)后自动关闭",
        "operationId": "qr_status_api_auth_qr_status_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "轮询二维码登录状态",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/sms/send": {
      "post": {
        "description": "## 业务说明\n向指定手机号发送短信验证码。\n\n## 使用流程\n1. 先调用 `/sms/init` 初始化\n2. 调用此接口发送验证码到手机\n3. 用户收到验证码后，调用 `/sms/verify` 验证\n\n## 注意事项\n- 手机号必须是已在教务系统注册的号码\n- 验证码有效期通常为5分钟\n- 发送间隔限制：60秒内只能发送一次\n- 如果频繁发送会返回429错误",
        "operationId": "sms_send_api_auth_sms_send_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "发送短信验证码",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/sms/verify": {
      "post": {
        "description": "## 业务说明\n验证短信验证码并完成登录。\n\n## 使用流程\n1. 用户输入收到的验证码\n2. 调用此接口验证\n3. 验证成功后即可访问需要认证的API\n\n## 注意事项\n- 验证码错误会返回401错误\n- 验证成功后会自动轮换session ID\n- 会设置新的session cookie",
        "operationId": "sms_verify_api_auth_sms_verify_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "验证短信验证码",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/status": {
      "get": {
        "description": "## 业务说明\n检查当前用户的认证状态，用于判断是否需要登录。\n\n## 使用场景\n- 页面加载时检查登录状态\n- 路由守卫判断是否需要跳转登录页\n- 定期检查session是否过期\n\n## 返回说明\n- `authenticated: true` - 已登录，可以访问受保护的API\n- `authenticated: false` - 未登录或session已过期\n- `state` - 当前认证状态（仅在已登录时返回）\n\n## 注意事项\n- 此接口不会抛出401错误，始终返回200\n- 即使没有cookie也会返回 `authenticated: false`",
        "operationId": "auth_status_api_auth_status_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "检查认证状态",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
...
}